+++ com.unity.visualeffectgraph/Editor/GraphView/Views/Properties/PropertyRM.cs
@@ -70,7 +70,7 @@ namespace UnityEditor.VFX.UI
             }
         }
 
-        IEnumerable<int> IPropertyRMProvider.filteredOutEnumerators { get { return null; } }
+        public virtual IEnumerable<int>  filteredOutEnumerators { get { return null; } }
 
         string IPropertyRMProvider.name
         {
@@ -104,7 +104,7 @@ namespace UnityEditor.VFX.UI
 
         static Texture2D[] m_IconStates;
 
-        public Label m_Label;
+        protected Label m_Label;
 
 
         public bool m_PropertyEnabled;
@@ -131,7 +131,7 @@ namespace UnityEditor.VFX.UI
                 UpdateIndeterminate();
             }
         }
-        public bool isDelayed { get; set; }
+        public virtual bool isDelayed { get; set; }
 
         protected bool hasChangeDelayed { get; set; }