+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -31,8 +31,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Ability to read attribute in spawn context through graph
 - Added save button to save only the current visual effect graph.
 - Added Degrees / Radians conversion subgraphs in samples
+- uint parameter can be seen as an enum.
 - New TransformVector4 operator
 - New GetTextureDimensions operator
+- Output Event context for scripting API event retrieval.
 
 ### Fixed
 - Moved VFX Event Tester Window visibility to Component Play Controls SceneView Window